+++ packages/react-utils/src/types/index.ts
@@ -1,1 +1,3 @@
-export type BooleanLike = any
+import type { HTMLAttributes, DetailedHTMLProps as _DetailedHTMLProps } from 'react'
+
+export type DetailedHTMLProps<T = HTMLDivElement> = _DetailedHTMLProps<HTMLAttributes<T>, T>