+++ packages/util-core/src/typings/deep.ts
@@ -1,0 +1,17 @@
+export type DeepReadonly<T> = {
+  readonly [P in keyof T]: T[P] extends object ? DeepReadonly<T[P]> : T[P]
+}
+
+export type DeepRequired<T> = {
+  [P in keyof T]-?: T[P] extends object ? DeepRequired<T[P]> : T[P]
+}
+
+export type DeepPartial<T> = {
+  [P in keyof T]?: T[P] extends object ? DeepPartial<T[P]> : T[P]
+}
+
+export type DeepReplace<T, K = unknown, V = unknown> = {
+  [P in keyof T]: K extends P ? V : DeepReplace<T[P], K, V>
+}
+
+export type DeepKeyof<T> = T extends object ? keyof T | DeepKeyof<T[keyof T]> : never