+++ packages/util-core/src/typings/util.ts
@@ -1,0 +1,14 @@
+import type { AtwillNumber, Key } from './atom'
+
+export type Awaitable<T> = T | Promise<T>
+export type Arrayable<T> = T | T[]
+
+export type Option<
+  L extends Key = 'label', V extends Key = 'value', C extends Key = 'children',
+> = {
+  [P in L]?: string
+} & {
+  [P in V]?: AtwillNumber
+} & {
+  [P in C]?: Option<L, V, C>[]
+}