+++ packages/util-core/src/util/index.ts
@@ -15,18 +15,24 @@ export function arange(x1: number, x2?: number, stp = 1, z: number[] = [], z0 = 
   return z
 }
 
-export type Deferred<T = void> = Promise<T> & { resolve: (value: T) => void; reject: Function }
+export type Deferred<T = void> = Promise<T> & {
+  resolve: (value: T) => void
+  reject: (value?: any) => void
+}
 
 export function createDeferred<T = void>(): Deferred<T> {
   let resolve: any, reject: any
+
   const promise = new Promise<any>((_resolve, _reject) => {
     resolve = _resolve
     reject = _reject
   }) as unknown as any
+
   promise.resolve = (v: any) => {
     resolve(v)
     return promise
   }
   promise.reject = reject
+
   return promise
 }