+++ packages/util-format/src/number/index.ts
@@ -1,17 +1,29 @@
-export interface PrefixZeroOptions {
-  type?: 'positive' | 'reverse'
-}
+import Bignumber from 'bignumber.js'
+
 /**
- * 数字位数不够时，进行前补零
- * @param number_ 数值
- * @param lh 长度
+ *  Any type that can be used where a numeric value is needed.
  */
-export function prefixZero(value: number | string, lh = 2, options: PrefixZeroOptions = {}) {
-  const { type = 'positive' } = options
-  value = String(integer(value))
-  if (value.length >= lh)
+export type Numeric = number | bigint
+
+/**
+ *  Any type that can be used where a big number is needed.
+ */
+export type Numberish = string | Numeric | { toString: (...args: any[]) => string }
+
+/**
+ * leading zeros
+ * @param number_
+ * @param lh
+ */
+export function zerofill(
+  value: Numberish,
+  n = 2,
+  type: 'positive' | 'reverse' = 'positive',
+) {
+  const _value = integer(value)
+  if (_value.length >= n)
     return value
-  const zero = '0'.repeat(lh - value.length)
+  const zero = '0'.repeat(n - _value.length)
   if (type === 'positive')
     return zero + value
   if (type === 'reverse')
@@ -19,25 +31,34 @@ export function prefixZero(value: number | string, lh = 2, options: PrefixZeroOp
   return ''
 }
 
-/**
- * 格式化为正整数
- * @param value 传入字符
- */
-export function integer(value: string | number) {
-  return 0 | +value
+export function numerfix(value: any) {
+  const _isNaN = isNaN(Number(value)) || value.toString() === 'NaN'
+  if (_isNaN)
+    console.warn(`numerfix(${value}): value is not the correct value. To ensure the normal operation of the program, it will be converted to zero`)
+  return (_isNaN) ? '0' : String(value)
 }
 
 /**
- * 保留 lh 位小数点
- * @param value 值
- * @param lh 小数点长度
+ * format as a positive integer
+ * @param value
  */
-export function keepDecimal(value: string | number, lh = 2) {
-  let [integer, decimal] = String(value).split('.')
+export function integer(value: Numberish) {
+  return new Bignumber(numerfix(value)).toFixed(0)
+}
+
+/**
+ * retain n decimal places
+ * @param value
+ * @param n
+ */
+export function decimal(value: Numberish, n = 2) {
+  let [integer, decimal] = numerfix(value).split('.')
+  if (n <= 0)
+    return integer
   if (!decimal)
     decimal = '0'
-  decimal = decimal.slice(0, lh)
-  decimal = decimal + '0'.repeat(lh - decimal.length)
+  decimal = decimal.slice(0, n)
+  decimal = decimal + '0'.repeat(n - decimal.length)
   return `${integer}.${decimal}`
 }
 
@@ -46,18 +67,20 @@ interface ThousandBitSeparatorOptions {
   decimal?: boolean
 }
 /**
- * 格式化数字千位分隔符
- * @param target 数值
- * @param unit 单位
+ * format number thousand separator
+ * @param target
+ * @param unit
  */
-export function thousandBitSeparator(target: number | string,
+export function thousandBitSeparator(
+  target: Numberish,
   unit = ',',
-  options: ThousandBitSeparatorOptions = {}) {
+  options: ThousandBitSeparatorOptions = {},
+) {
   options.integer = options.integer ?? true
   options.decimal = options.decimal ?? true
   const exp = /(\d)(?=(\d{3})+$)/ig
   const replace = (v: string) => v.replace(exp, `$1${unit || ''}`)
-  let [integer = '0', decimal = ''] = String(target).split('.')
+  let [integer = '0', decimal = ''] = numerfix(target).split('.')
   if (options.integer)
     integer = replace(integer)
   if (options.decimal)
@@ -66,15 +89,17 @@ export function thousandBitSeparator(target: number | string,
 }
 
 /**
- * 计算百分比（仅返回整数）
- * @param total 总数
- * @param count 数量
- * @returns
+ * calculate percentage
+ * @param total
+ * @param count
  */
-export function percentage(total: number, count: number) {
-  if (+total === 0)
-    return 0
-  if (+count === 0)
-    return 0
-  return integer((count / total) * 100)
+export function percentage(total: Numberish, count: Numberish, decimals = 0) {
+  const _total = new Bignumber(numerfix(total))
+  const _count = new Bignumber(numerfix(count))
+  if (_total.eq(0))
+    return '0'
+  if (_count.eq(0))
+    return '0'
+
+  return _count.dividedBy(_total).multipliedBy(100).toFixed(decimals, Bignumber.ROUND_DOWN)
 }